requirements.txt
langgraph>=0.2
langchain>=0.3
langchain-anthropic
langchain-openai
langchain-groq
langchain-community
faiss-cpu
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2.0
httpx
